# server/dist/generated/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "MessageResponse": {
                "properties": {
                    "messageId": {
                        "type": "string"
                    },
                    "conversationId": {
                        "type": "string"
                    },
                    "senderId": {
                        "type": "string"
                    },
                    "receiverId": {
                        "type": "string"
                    },
                    "senderUsername": {
                        "type": "string"
                    },
                    "senderDisplayName": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "readBy": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "sentAt": {
                        "type": "string"
                    }
                },
                "required": [
                    "messageId",
                    "conversationId",
                    "senderId",
                    "receiverId",
                    "senderUsername",
                    "senderDisplayName",
                    "content",
                    "readBy",
                    "sentAt"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "chat-poc-server",
        "version": "1.0.0",
        "description": "ChatPOC backend — Express + TSOA + TypeORM + Socket.IO",
        "contact": {}
    },
    "paths": {
...
        "/messages/{receiverId}": {
            "get": {
                "operationId": "GetMessages",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MessageResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Get paginated message history with a specific user.",
                "tags": [
                    "Messages"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "receiverId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        },
        "/messages/{receiverId}/search": {
            "get": {
                "operationId": "SearchMessages",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MessageResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Full-text search messages with a specific user.",
                "tags": [
                    "Messages"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "receiverId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/messages/{messageId}/read": {
            "patch": {
                "operationId": "MarkRead",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "description": "Mark a message as read.",
                "tags": [
                    "Messages"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "messageId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}